# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SheetCounts {
  pzm: number;
  psm: number;
  pstl: number;
  vstl: number;
  dozh: number;
  date: string;
}

export interface TelegramChannel {
  id: number;
  name: string;
  chatId: string;
  createdAt: string;
}

export interface TelegramChannelInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  chatId: string;
  /** @minLength 1 */
  botToken: string;
}

export interface TelegramChannelUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  chatId?: string;
  botToken?: string;
}

export interface TelegramSendInput {
  channelId: number;
  /** @minLength 1 */
  text: string;
}

export interface TelegramSendResult {
  ok: boolean;
  /** @nullable */
  messageId?: number | null;
}

export interface ReportStateData {
  /** @minimum 0 */
  pzm: number;
  /** @minimum 0 */
  psm: number;
  /** @minimum 0 */
  pstl: number;
  /** @minimum 0 */
  vstl: number;
  /** @minimum 0 */
  dozh: number;
  trafikPlan: string;
  trafikCurrent: string;
  /** @minimum 0 */
  kz: number;
  prihod: string;
  uhod: string;
  /** @minimum 0 */
  postupleniya: number;
  /** @minimum 0 */
  itogoFact: number;
  /** @minimum 0 */
  itogoPlan: number;
  /** @minimum 0 */
  vypolnenie: number;
  financesEnabled: boolean;
}

export interface ReportSignature {
  /** @minLength 1 */
  tag1: string;
  /** @minLength 1 */
  tag2: string;
  /** @minLength 1 */
  mention: string;
}

export interface UserReportStateInput {
  state: ReportStateData;
  signature: ReportSignature;
}

export interface AutoReportScheduleInput {
  /** @minimum 1 */
  channelId: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  intervalMinutes: number;
  /**
     * @minimum 0
     * @maximum 2
     */
  reportType: number;
  isActive: boolean;
}

export interface AutoReportScheduleUpdate {
  /** @minimum 1 */
  channelId?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  intervalMinutes?: number;
  /**
     * @minimum 0
     * @maximum 2
     */
  reportType?: number;
  isActive?: boolean;
}

export interface AutoReportSchedule {
  id: number;
  channelId: number;
  intervalMinutes: number;
  reportType: number;
  isActive: boolean;
  /** @nullable */
  lastSentAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export type GetSheetCountsParams = {
/**
 * Manager name to filter by (column K). Normalized match — spaces and case ignored.
 */
name?: string;
};