# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: telegram
    description: Telegram channel management and message delivery
  - name: sheets
    description: Google Sheets data sync
  - name: auto-reports
    description: Scheduled Telegram report delivery
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /sheets/counts:
    get:
      operationId: getSheetCounts
      tags: [sheets]
      summary: Count today's touches per metric from Google Sheets
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Manager name to filter by (column K). Normalized match — spaces and case ignored.
      responses:
        "200":
          description: Today's metric counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetCounts"
        "401":
          description: Unauthorized
        "502":
          description: Google Sheets API error

  /telegram/channels:
    get:
      operationId: listTelegramChannels
      tags: [telegram]
      summary: List the authenticated user's Telegram channels
      responses:
        "200":
          description: List of channels (bot tokens are never returned)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelegramChannel"
        "401":
          description: Unauthorized
    post:
      operationId: createTelegramChannel
      tags: [telegram]
      summary: Add a new Telegram channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramChannelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramChannel"
        "401":
          description: Unauthorized

  /telegram/channels/{id}:
    put:
      operationId: updateTelegramChannel
      tags: [telegram]
      summary: Update a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramChannelUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramChannel"
        "401":
          description: Unauthorized
        "404":
          description: Not found
    delete:
      operationId: deleteTelegramChannel
      tags: [telegram]
      summary: Delete a channel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /telegram/send:
    post:
      operationId: sendTelegramMessage
      tags: [telegram]
      summary: Send a message to a channel via the server (bot token stays server-side)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramSendInput"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramSendResult"
        "401":
          description: Unauthorized
        "404":
          description: Channel not found
        "502":
          description: Telegram API error

  /state:
    post:
      operationId: saveUserReportState
      tags: [auto-reports]
      summary: Save the authenticated user's current report state for scheduled delivery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserReportStateInput"
      responses:
        "204":
          description: State saved
        "401":
          description: Unauthorized

  /auto-reports:
    get:
      operationId: getAutoReportSchedule
      tags: [auto-reports]
      summary: Get the authenticated user's scheduled report configuration
      responses:
        "200":
          description: Current schedule, or null if it has not been configured
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AutoReportSchedule"
                  - type: "null"
        "401":
          description: Unauthorized
    post:
      operationId: createAutoReportSchedule
      tags: [auto-reports]
      summary: Create or activate the authenticated user's scheduled report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoReportScheduleInput"
      responses:
        "200":
          description: Schedule saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoReportSchedule"
        "400":
          description: Invalid input or channel
        "401":
          description: Unauthorized
    patch:
      operationId: updateAutoReportSchedule
      tags: [auto-reports]
      summary: Update the authenticated user's scheduled report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoReportScheduleUpdate"
      responses:
        "200":
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoReportSchedule"
        "401":
          description: Unauthorized
        "404":
          description: Schedule not found
    delete:
      operationId: deleteAutoReportSchedule
      tags: [auto-reports]
      summary: Deactivate the authenticated user's scheduled report
      responses:
        "204":
          description: Schedule deactivated
        "401":
          description: Unauthorized
        "404":
          description: Schedule not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SheetCounts:
      type: object
      properties:
        pzm:
          type: number
        psm:
          type: number
        pstl:
          type: number
        vstl:
          type: number
        dozh:
          type: number
        date:
          type: string
      required:
        - pzm
        - psm
        - pstl
        - vstl
        - dozh
        - date

    TelegramChannel:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        chatId:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - chatId
        - createdAt

    TelegramChannelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        chatId:
          type: string
          minLength: 1
        botToken:
          type: string
          minLength: 1
      required:
        - name
        - chatId
        - botToken

    TelegramChannelUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        chatId:
          type: string
          minLength: 1
        botToken:
          type: string

    TelegramSendInput:
      type: object
      properties:
        channelId:
          type: number
        text:
          type: string
          minLength: 1
      required:
        - channelId
        - text

    TelegramSendResult:
      type: object
      properties:
        ok:
          type: boolean
        messageId:
          type: ["number", "null"]
      required:
        - ok

    ReportStateData:
      type: object
      properties:
        pzm: { type: number, minimum: 0 }
        psm: { type: number, minimum: 0 }
        pstl: { type: number, minimum: 0 }
        vstl: { type: number, minimum: 0 }
        dozh: { type: number, minimum: 0 }
        trafikPlan: { type: string }
        trafikCurrent: { type: string }
        kz: { type: number, minimum: 0 }
        prihod: { type: string }
        uhod: { type: string }
        postupleniya: { type: number, minimum: 0 }
        itogoFact: { type: number, minimum: 0 }
        itogoPlan: { type: number, minimum: 0 }
        vypolnenie: { type: number, minimum: 0 }
        financesEnabled: { type: boolean }
      required: [pzm, psm, pstl, vstl, dozh, trafikPlan, trafikCurrent, kz, prihod, uhod, postupleniya, itogoFact, itogoPlan, vypolnenie, financesEnabled]

    ReportSignature:
      type: object
      properties:
        tag1: { type: string, minLength: 1 }
        tag2: { type: string, minLength: 1 }
        mention: { type: string, minLength: 1 }
      required: [tag1, tag2, mention]

    UserReportStateInput:
      type: object
      properties:
        state:
          $ref: "#/components/schemas/ReportStateData"
        signature:
          $ref: "#/components/schemas/ReportSignature"
      required: [state, signature]

    AutoReportScheduleInput:
      type: object
      properties:
        channelId: { type: number, minimum: 1 }
        intervalMinutes: { type: number, minimum: 1, maximum: 10 }
        reportType: { type: number, minimum: 0, maximum: 2 }
        isActive: { type: boolean }
      required: [channelId, intervalMinutes, reportType, isActive]

    AutoReportScheduleUpdate:
      type: object
      properties:
        channelId: { type: number, minimum: 1 }
        intervalMinutes: { type: number, minimum: 1, maximum: 10 }
        reportType: { type: number, minimum: 0, maximum: 2 }
        isActive: { type: boolean }

    AutoReportSchedule:
      type: object
      properties:
        id: { type: number }
        channelId: { type: number }
        intervalMinutes: { type: number }
        reportType: { type: number }
        isActive: { type: boolean }
        lastSentAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, channelId, intervalMinutes, reportType, isActive, lastSentAt, createdAt, updatedAt]