# lib/api-zod/src/generated/types/autoReportSchedule.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoReportSchedule {
  id: number;
  channelId: number;
  intervalMinutes: number;
  reportType: number;
  isActive: boolean;
  /** @nullable */
  lastSentAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}