# lib/api-zod/src/generated/types/autoReportScheduleUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoReportScheduleUpdate {
  /** @minimum 1 */
  channelId?: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  intervalMinutes?: number;
  /**
     * @minimum 0
     * @maximum 2
     */
  reportType?: number;
  isActive?: boolean;
}