# lib/api-zod/src/generated/types/reportSignature.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportSignature {
  /** @minLength 1 */
  tag1: string;
  /** @minLength 1 */
  tag2: string;
  /** @minLength 1 */
  mention: string;
}