# lib/api-zod/src/generated/types/reportStateData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportStateData {
  /** @minimum 0 */
  pzm: number;
  /** @minimum 0 */
  psm: number;
  /** @minimum 0 */
  pstl: number;
  /** @minimum 0 */
  vstl: number;
  /** @minimum 0 */
  dozh: number;
  trafikPlan: string;
  trafikCurrent: string;
  /** @minimum 0 */
  kz: number;
  prihod: string;
  uhod: string;
  /** @minimum 0 */
  postupleniya: number;
  /** @minimum 0 */
  itogoFact: number;
  /** @minimum 0 */
  itogoPlan: number;
  /** @minimum 0 */
  vypolnenie: number;
  financesEnabled: boolean;
}