# lib/api-zod/src/generated/types/userReportStateInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportSignature } from './reportSignature';
import type { ReportStateData } from './reportStateData';

export interface UserReportStateInput {
  state: ReportStateData;
  signature: ReportSignature;
}